requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pandas>=2.1
pdfplumber>=0.10
matplotlib>=3.8
python-multipart>=0.0.6
pdfminer.six>=20221105
gunicorn>=21.2

